Add void to function declarations for EVP_EncryptInit check
gcc 11.1.1 complains about not having strict function prototypes
(-Wstrict-prototypes) when attempting to test for linkage with openssl.

This warning was added in PR #284. Why it started acting up now is
unknown. It may be that gcc 11.1.1 have changed how and when they emit
this warning.

Adding void to the function declarations seems to silence this warning
and allow the test to run to completion.